left-sidenav clicking closes bar, and moved component to pages
