+++ src/app/deferred.routes.ts
@@ -4,7 +4,7 @@ import { provideStore } from "@ngrx/store";
 import { MainComponent } from "./pages/main/main.component";
 import { RightTabNavComponent } from "./pages/right-tab-nav/right-tab-nav.component";
 import { FileIconModule } from "./shared/components/file-icon/file-icon.module";
-import { SidenavComponent } from "./shared/components/sidenav/sidenav.component";
+import { SidenavComponent } from "./pages/sidenav/sidenav.component";
 import { MeshNormalMaterialService } from "./shared/services/3d-managers/mesh-normal-material.service";
 import { AnnotationBuilderService } from "./shared/services/annotation-builder.service";
 import { CanvasService } from "./shared/services/canvas.service";
