+++ src/app/pages/sidenav/sidenav.component.html
@@ -1,4 +1,4 @@
-<mat-list (click)="buttonClicked.emit()">
+<mat-list>
   <button mat-list-item (click)="dialogOpener.open()">Open...</button>
   <button mat-list-item (click)="dialogOpener.import()">Import...</button>
   <button mat-list-item (click)="dialogOpener.save()">Zip...</button>
