+++ src/app/pages/sidenav/sidenav.component.ts
@@ -1,9 +1,9 @@
 import { AsyncPipe } from '@angular/common';
-import { ChangeDetectionStrategy, Component, EventEmitter, Output, inject } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { MatListModule } from '@angular/material/list';
-import { DialogOpenerService } from '../../services/dialog-opener.service';
-import { ServiceWorkerService } from '../../services/service-worker.service';
-import { MAPPER_VERSION } from '../../tokens/version.token';
+import { DialogOpenerService } from '../../shared/services/dialog-opener.service';
+import { ServiceWorkerService } from '../../shared/services/service-worker.service';
+import { MAPPER_VERSION } from '../../shared/tokens/version.token';
 
 @Component({
   selector: 'mapper-sidenav',
@@ -18,9 +18,6 @@ export class SidenavComponent {
   readonly mapperVersion = inject(MAPPER_VERSION).version;
   readonly dialogOpener = inject(DialogOpenerService);
   readonly #serviceWorker = inject(ServiceWorkerService);
-
-  @Output()
-  readonly buttonClicked = new EventEmitter<void>();
 
   checkForUpdate() {
     this.#serviceWorker.checkForUpdate().subscribe();
